+++ client/app/components/project/project.component.ts
@@ -92,11 +92,19 @@ export class ProjectComponent {
   deleteProject() {
     let choice = prompt('Enter the projects the title of the project you wish to delete');
     if (choice === this.project.title) {
-      this.projectService.deleteProject(this.id)
-        .subscribe(
-          data => this.router.navigateByUrl('/'),
-          err => err
-        )
+      if(this.memberType === '') {
+        this.projectService.deleteProject(this.id)
+          .subscribe(
+            data => this.router.navigateByUrl('/'),
+            err => err
+          )
+      } else {
+        this.projectService.teamDeleteProject(this.project.Profile.id, this.id)
+          .subscribe(
+            data => this.router.navigateByUrl('/'),
+            err => err
+          )
+      }  
     }
   }
 
@@ -262,7 +270,7 @@ export class ProjectComponent {
       this.determineOpenSource(this.project.contribute);
     }
     this.project[type] = input[type]
-    //this.project.descripiton = input.descripiton;
+
     if(this.memberType === '') {
       this.projectService.editDescription(this.id, input)
         .subscribe(
@@ -270,7 +278,6 @@ export class ProjectComponent {
           err => err
         )
     } else {
-      console.log(this.project)
       this.projectService.teamEditDescription(this.project.Profile.id, this.id, input)
         .subscribe(
           data => this.editingProject(type),
